+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add link to executables to installation manual
 
+### Fixed
+- Incorrect autoselection of suffixed options
+
 ## [1.3.0] - 2020-07-21
 ### Changed 
 - History does not contain duplicate items